+++ parents.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * a dump of the module chain for this module for static code analysers
+ * taken from the ide helper .phpstorm.meta.php/oxid.meta.php
+ */
+
+
+namespace OxidProfessionalServices\PasswordPolicy\Core {
+    class PasswordPolicyViewConfig_parent extends \OxidEsales\Eshop\Core\ViewConfig
+    {
+
+    }
+}
+
+
+namespace OxidProfessionalServices\PasswordPolicy\Core {
+    class PasswordPolicyValidator_parent extends \OxidEsales\Eshop\Core\InputValidator
+    {
+
+    }
+}