+++ src/Core/PasswordPolicyValidator.php
@@ -1,8 +1,29 @@
 <?php
 
+/**
+ * This file is part of OXID Professional Services Password Policy module.
+ *
+ * OXID Professional Services Password Policy module is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * OXID Professional Services Password Policy module is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with OXID Professional Services Password Policy module.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author        OXID Professional services
+ * @link          http://www.oxid-esales.com
+ * @copyright (C) OXID eSales AG 2003-2019
+ */
+
+declare(strict_types=1);
 
 namespace OxidProfessionalServices\PasswordPolicy\Core;
-
 
 use OxidEsales\Eshop\Application\Model\User;
 use OxidEsales\Eshop\Core\Exception\InputException;
@@ -27,9 +48,9 @@ class PasswordPolicyValidator extends PasswordPolicyValidator_parent
         return parent::checkPassword($user, $newPassword, $confirmationPassword, $shouldCheckPasswordLength);
     }
 
-    public function getModuleSettings()
+    public function getModuleSettings(): PasswordPolicyConfig
     {
-        return Registry::get(PasswordPolicyConfig::class)->getModuleSettings();
+        return Registry::get(PasswordPolicyConfig::class);
     }
 
     /**
@@ -45,31 +66,31 @@ class PasswordPolicyValidator extends PasswordPolicyValidator_parent
         $iPasswordLength = mb_strlen($sPassword, 'UTF-8');
 
         // Load module settings
-        $aSettings = $this->getModuleSettings();
+        $settings = $this->getModuleSettings();
 
         // Validate password according to settings params
-        if ($iPasswordLength < $aSettings['iMinPasswordLength']) {
+        if ($iPasswordLength < $this->getPasswordLength()) {
             $sError = 'ERROR_MESSAGE_PASSWORD_TOO_SHORT';
         }
 
-        if ($iPasswordLength > $aSettings['iMaxPasswordLength']) {
+        if ($iPasswordLength > $settings->getMaxPasswordLength()) {
             $sError = 'OXPS_PASSWORDPOLICY_PASSWORDSTRENGTH_ERROR_TOOLONG';
         }
 
-        if (!empty($aSettings['aPasswordRequirements']['digits']) and !preg_match('(\d+)', $sPassword)) {
+        if ($settings->getPasswordNeedDigits() and !preg_match('(\d+)', $sPassword)) {
             $sError = 'OXPS_PASSWORDPOLICY_PASSWORDSTRENGTH_ERROR_REQUIRESDIGITS';
         }
 
-        if (!empty($aSettings['aPasswordRequirements']['capital']) and !preg_match('(\p{Lu}+)', $sPassword)) {
+        if ($settings->getPasswordNeedUpperCase() and !preg_match('(\p{Lu}+)', $sPassword)) {
             $sError = 'OXPS_PASSWORDPOLICY_PASSWORDSTRENGTH_ERROR_REQUIRESCAPITAL';
         }
 
-        if (!empty($aSettings['aPasswordRequirements']['lower']) and !preg_match('(\p{Ll}+)', $sPassword)) {
+        if ($settings->getPasswordNeedLowerCase() and !preg_match('(\p{Ll}+)', $sPassword)) {
             $sError = 'OXPS_PASSWORDPOLICY_PASSWORDSTRENGTH_ERROR_REQUIRESLOWER';
         }
 
         if (
-            !empty($aSettings['aPasswordRequirements']['special']) and
+            $settings->getPasswordNeedSpecialCharacter() and
             !preg_match('([\.,_@\~\(\)\!\#\$%\^\&\*\+=\-\\\/|:;`]+)', $sPassword)
         ) {
             $sError = 'OXPS_PASSWORDPOLICY_PASSWORDSTRENGTH_ERROR_REQUIRESSPECIAL';
@@ -86,4 +107,13 @@ class PasswordPolicyValidator extends PasswordPolicyValidator_parent
         return $res;
     }
 
+    /**
+     * Min length of password.
+     *
+     * @return int
+     */
+    public function getPasswordLength()
+    {
+        return Registry::getConfig()->getConfigParam(PasswordPolicyConfig::SettingMinPasswordLength, 8);
+    }
 }